+++ app/app/game_logic/local_func.py
@@ -1,0 +1,30 @@
+from game_logic.game_loop import *
+
+def process_game_events():
+    board = game_state['board']
+    board_container = game_state['container']
+    player_color = game_state['player_color']
+    
+    if board.is_checkmate() or board.is_stalemate():
+        print("Partie terminée.")
+        print(board, "\n")
+        if board.is_checkmate():
+            print("Échec et mat ! Le joueur", "Blanc" if board.turn == chess.WHITE else "Noir", "a gagné.")
+        elif board.is_stalemate():
+            print("Partie nulle par pat.")
+        # Gérer d'autres fins de partie si nécessaire
+        return
+
+    event = get_next_event()
+    if event:
+        handle_event(event)
+
+    board_container.after(50, process_game_events)
+
+def handle_local_move_event():
+    """
+    Gère la logique de lecture des événements UART pour un coup local.
+    """
+    event = get_next_event()
+    if event:
+        handle_event(event)
