--- app/app/lichess_api/lichess_api.py
@@ -1,6 +1,0 @@
-import os
-import berserk
-
-def connect(token: str):
-    session = berserk.TokenSession(token)
-    return berserk.Client(session=session)
